Delete pose.py and implement its functionality in controller.py
